requirements.txt
attrs==21.2.0
certifi==2020.12.5
click==8.0.1
cycler==0.10.0
flasgger==0.9.5
Flask==2.0.0
gunicorn==20.1.0
itsdangerous==2.0.1
jedi==0.17.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
mistune==0.8.4
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
pyzmq==20.0.0
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
sklearn==0.0
streamlit==0.82.0
threadpoolctl==2.1.0
Werkzeug==2.0.1
wincertstore==0.2
